renamed train & predict -> xxx_application
