+++ hdnnpy/cli/__init__.py
@@ -4,7 +4,4 @@ __all__ = [
     'main',
     ]
 
-from hdnnpy.cli.main import HDNNPApplication
-
-
-main = HDNNPApplication.launch_instance
+from hdnnpy.cli.main import main
