+++ hdnnpy/cli/main.py
@@ -7,8 +7,8 @@ import sys
 from traitlets import Unicode
 from traitlets.config import Application
 
-from hdnnpy.cli.predict import PredictionApplication
-from hdnnpy.cli.train import TrainingApplication
+from hdnnpy.cli.prediction_application import PredictionApplication
+from hdnnpy.cli.training_application import TrainingApplication
 from hdnnpy.utils import MPI
 
 
@@ -22,13 +22,10 @@ class HDNNPApplication(Application):
         'train': (TrainingApplication, TrainingApplication.description),
         }
 
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
     def initialize(self, argv=None):
         if MPI.rank != 0:
             sys.stdout = Path(os.devnull).open('w')
         super().initialize(argv)
 
-    def start(self):
-        super().start()
+
+main = HDNNPApplication.launch_instance
